move jormungandr settings to a separate module and impl Deserialize
